backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
sqlalchemy==2.0.20
alembic==1.12.0
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.1.1
pydantic-settings==2.0.3
aiohttp==3.8.5
aioredis==2.0.1
python-memcached==1.59
geopy==2.4.0
pandas==2.1.1
python-dateutil==2.8.2
tenacity==8.2.3
httpx==0.24.1
python-slugify==8.0.1
python-multipart==0.0.6
